Remove renderer limitation on benchmark scores
Some clients like RetroGecko can specify renderers that aren't part of the vanilla game. Other clients might be using DXVK and might want to expose that to benchmark scores. For example, this field could be `Vulkan` or `D3D11`.

This commit removes the check restricting the renderer to just `OpenGL` and `DirectX` to allow for these types of clients.